+++ spec/sisimai/bite/email/private-amavis_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+require './spec/sisimai/bite/email/code'
+enginename = 'Amavis'
+isexpected = [
+  { 'n' => '01001', 'r' => /userunknown/ },
+  { 'n' => '01002', 'r' => /userunknown/ },
+]
+Sisimai::Bite::Email::Code.maketest(enginename, isexpected, true)
+